--- client/src/App.css
@@ -1,3 +1,0 @@
-.main {
-  margin-top: 300px;
-}
