+++ client/src/App.tsx
@@ -5,7 +5,7 @@ import "./App.css";
 function App() {
   return (
     <>
-      <Navbar />
+      {/* <Navbar /> */}
       <div className="main">
         <Outlet />
       </div>
