+++ client/src/Pages/AboutPage/AboutPage.tsx
@@ -1,23 +1,42 @@
-import { useEffect, useState } from "react";
+import { act, useEffect, useState } from "react";
 import { AnimatePresence, motion } from "framer-motion";
 import "./AboutPage.css";
 const AboutPage = () => {
-    const [activeSection, setActiveSection] = useState<number>(0);
+    const [activeSection, setActiveSection] = useState<number>(-1);
+    const [persisting, setPersisting] = useState("persisting-fake");
 
     // Handle scroll to determine which section should be active
     const handleScroll = () => {
         const sections = document.querySelectorAll<HTMLElement>(".section");
         const scrollPosition = window.scrollY + window.innerHeight / 2;
-
+        // console.log(scrollPosition)
         sections.forEach((section, index) => {
             const sectionTop = section.offsetTop;
             const sectionHeight = section.offsetHeight;
+            // console.log(index, sectionTop);
+            // console.log(index, sectionHeight);
 
             if (
                 scrollPosition >= sectionTop &&
                 scrollPosition < sectionTop + sectionHeight
             ) {
                 setActiveSection(index);
+            }
+
+            if (
+                index == 1 &&
+                scrollPosition >= sectionTop - sectionHeight / 2 &&
+                persisting != "persisting"
+            ) {
+                console.log("start persisting");
+                setPersisting("persisting");
+                // console.log(scrollPosition);
+                // console.log(sectionTop - sectionHeight/2);
+            } else if (
+                index == 1 &&
+                scrollPosition <= sectionTop - sectionHeight / 2
+            ) {
+                setPersisting("persisting-fake");
             }
         });
     };
@@ -28,9 +47,12 @@ const AboutPage = () => {
             window.removeEventListener("scroll", handleScroll);
         };
     }, []);
+    useEffect(() => {
+        // console.log("activeSEction", activeSection);
+    }, [activeSection]);
 
     return (
-        <div style={{ height: "300vh" }}>
+        <div style={{ height: "300vh", border: "2px solid black" }}>
             {/* Section 1 */}
             <motion.section
                 className="section"
@@ -40,7 +62,7 @@ const AboutPage = () => {
                 <AnimatePresence>
                     {activeSection === 0 && (
                         <motion.div
-                            className="persisting"
+                            className={persisting}
                             animate={{ x: 100, opacity: 1 }}
                             transition={{ duration: 0.5 }}
                             initial={{ opacity: 0 }}
@@ -113,16 +135,6 @@ const AboutPage = () => {
                 </motion.section>
             </AnimatePresence>
         </div>
-        // <div>
-        //     {/* <h1>About Page</h1> */}
-        //     {/* <MotionTesting></MotionTesting> */}
-        //     {/* <Variant></Variant> */}
-        //     <Envelope></Envelope>
-        //     <br></br>
-        //     <Envelope></Envelope>
-        //     <br></br>
-        //     <Envelope></Envelope>
-        // </div>
     );
 };
