+++ client/src/Pages/HomePage/HomePage copy.css
@@ -1,0 +1,82 @@
+.home-page {
+    padding-bottom: 200px;
+    padding-right: 50px;
+}
+
+.my-image {
+    width: 100%;
+    height: 100%;
+    object-fit: cover;
+}
+
+.image-container {
+    width: 30vw; /* Set the width you want */
+    height: 70vh; /* Set the height you want */
+    overflow: hidden; /* Hide any overflow */
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    border-radius: 50% 5%;
+    padding-left: 50px;
+}
+
+.home-front-container {
+    display: flex;
+    flex-direction: row;
+    justify-content: flex-start;
+    gap: 100px;
+    /* border:2px solid black; */
+}
+
+.fade-in {
+    opacity: 1;
+    transition: opacity 1s ease;
+}
+.fade-out {
+    opacity: 0;
+    transition: opacity 1s ease;
+}
+
+.home-intro {
+    width: 70%;
+    border-bottom: 2px solid black;
+    /* margin-right: 100px; */
+    padding-bottom: 15px;
+}
+
+.test {
+    width: 100%;
+}
+
+.content-title {
+    display: flex;
+    flex-direction: row;
+    align-items: baseline;
+    justify-content: space-between;
+}
+
+.content-skills {
+    display: flex;
+    flex-direction: row;
+    gap: 30px;
+    align-items: center;
+    padding-bottom: 0px;
+}
+
+.content-skills-list {
+    width: 100%;
+    display: flex;
+    flex-direction: row;
+    list-style-type: none;
+    margin: 0;
+    padding: 0;
+    justify-content: flex-start;
+    gap: 30px;
+}
+
+.experiences-container {
+    margin-left: 50px;
+    margin-right: 200px;
+    margin-top: 50px;
+    /* display: none; */
+}
